services/mqtt-bridge/requirements.txt
paho-mqtt==1.6.1
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
